packages/agents-py/pyproject.toml
[project]
name = "aegis-agents"
version = "0.1.0"
description = "AEGIS Protocol - AI Sentinel Agents (CrewAI + FastAPI)"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.14.0",
    "anthropic>=0.40.0",
    "web3>=7.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["aegis"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
